AppEnvironment: relocate to application directory
