+++ xbmc/CMakeLists.txt
@@ -2,7 +2,6 @@ set(SOURCES Application.cpp
             AppInboundProtocol.cpp
             ApplicationPlayer.cpp
             ApplicationStackHelper.cpp
-            AppEnvironment.cpp
             AppParamParser.cpp
             AppParams.cpp
             AutoSwitch.cpp
@@ -39,8 +38,7 @@ set(SOURCES Application.cpp
             Util.cpp
             XBDateTime.cpp)
 
-set(HEADERS AppEnvironment.h
-            AppParamParser.h
+set(HEADERS AppParamParser.h
             AppParams.h
             Application.h
             AppInboundProtocol.h
