+++ xbmc/application/AppEnvironment.cpp
