+++ xbmc/application/AppEnvironment.h
