+++ xbmc/application/CMakeLists.txt
@@ -1,11 +1,13 @@
-set(SOURCES ApplicationActionListeners.cpp
+set(SOURCES AppEnvironment.cpp
+            ApplicationActionListeners.cpp
             ApplicationPlayerCallback.cpp
             ApplicationPowerHandling.cpp
             ApplicationSettingsHandling.cpp
             ApplicationSkinHandling.cpp
             ApplicationVolumeHandling.cpp)
 
-set(HEADERS ApplicationActionListeners.h
+set(HEADERS AppEnvironment.h
+            ApplicationActionListeners.h
             ApplicationPlayerCallback.h
             ApplicationPowerHandling.h
             ApplicationSettingsHandling.h
