+++ xbmc/platform/darwin/tvos/XBMCController.mm
@@ -8,12 +8,12 @@
 
 #import "platform/darwin/tvos/XBMCController.h"
 
-#include "AppEnvironment.h"
 #include "AppParams.h"
 #include "Application.h"
 #include "CompileInfo.h"
 #include "FileItem.h"
 #include "ServiceBroker.h"
+#include "application/AppEnvironment.h"
 #include "cores/AudioEngine/Interfaces/AE.h"
 #include "guilib/GUIComponent.h"
 #include "guilib/GUIWindowManager.h"
